Bump Version For Release With Leak Timer Fix
Summary:
I'm still seeing the issue #2

```
Warning: setState(...): Can only update a mounted or mounting component. This usually means you called setState() on an unmounted component. This is a no-op. Please check the code for the Pulse component.
```

I noticed that the timer leak #3 is not released via npm. This bumps
the version in order to make a release easy. :)

Thanks for creating this component!